+++ demos/index.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-require '../vendor/autoload.php';
+namespace Atk4\Filestore\Demos;
 
 use Atk4\Filestore\Field\File;
 use Atk4\Filestore\Helper;
@@ -10,24 +10,10 @@ use Atk4\Ui\Callback;
 use Atk4\Ui\Columns;
 use Atk4\Ui\Form;
 use Atk4\Ui\JsExpression;
-use League\Flysystem\Adapter\Local;
 use League\Flysystem\Filesystem;
 
-class Friend extends \Atk4\Data\Model
-{
-    public $table = 'friend';
-
-    public $filesystem;
-
-    protected function init(): void
-    {
-        parent::init();
-
-        $this->addField('name');                                                     // friend's name
-        $this->addField('file', new File($this->filesystem));  // storing file here
-        $this->addField('file2', new File($this->filesystem)); // storing file here
-    }
-}
+require __DIR__ . '/../vendor/autoload.php';
+require __DIR__ . '/_includes/Friend.php';
 
 // specify folder where files will be actually stored
 $adapter = new \League\Flysystem\Local\LocalFilesystemAdapter(__DIR__ . '/localfiles');
@@ -38,16 +24,16 @@ $app = new \Atk4\Ui\App('Filestore Demo');
 $app->initLayout([\Atk4\Ui\Layout\Centered::class]);
 
 // init db
-$db_file = __DIR__ . "/filestore.db";
+$db_file = __DIR__ . '/filestore.db';
 $db_file_exists = file_exists($db_file);
 // change this as needed
 $app->db = Atk4\Data\Persistence::connect('sqlite:' . $db_file);
 /*
 if (!$db_file_exists) {
-    (new \Atk4\Schema\Migration(new \Friend($app->db, ['filesystem' => $filesystem,])))
+    (new \Atk4\Schema\Migrator(new \Friend($app->db, ['filesystem' => $filesystem,])))
         ->dropIfExists()
         ->create();
-    (new \Atk4\Schema\Migration(new \Atk4\Filestore\Model\File($app->db)))
+    (new \Atk4\Schema\Migrator(new \Atk4\Filestore\Model\File($app->db)))
         ->dropIfExists()
         ->create();
 }
